+++ src/ui/about.rs
@@ -93,16 +93,19 @@ impl SimpleComponent for AboutDialog {
                 "<p>Fixed</p>",
 
                 "<ul>",
-                    "<li>Added deletion of old patch files</li>",
-                    "<li>Added telemetry disabling state support</li>",
-                    "<li>Added Discord RPC icons updating</li>",
+                    "<li>Fixed telemetry disabling</li>",
                 "</ul>",
 
                 "<p>Changed</p>",
 
                 "<ul>",
-                    "<li>Replaced xlua patch by \"disable mhypbase\" option</li>",
-                    "<li>Returned back old background file path</li>",
+                    "<li>Updated Italian</li>",
+                    "<li>Updated Hungarian</li>",
+                    "<li>Updated Japanese</li>",
+                    "<li>Updated Indonesian</li>",
+                    "<li>Updated Spanish</li>",
+                    "<li>Updated Turkish</li>",
+                    "<li>Updated Chinese</li>",
                 "</ul>",
             ].join("\n"),